Initial sketch for importing compute images vs building them
This includes variables to define the source path of the images and
a place holder var for conditional tests. Ideally it should be a
flag that selects import vs build.

A modified build file to select a new role for import.

The new import role that replaces the build steps and everything
before the build with a simple import.